+++ cs4rsa_core/Dialogs/Implements/AutoSortSubjectLoadViewModel.cs
@@ -24,7 +24,7 @@ namespace cs4rsa_core.Dialogs.Implements
             IEnumerable<string> courseIds = ProgramSubjectModels.Select(item => item.CourseId);
             foreach (string courseId in courseIds)
             {
-                Subject subject = await _subjectCrawler.Crawl(ushort.Parse(courseId));
+                Subject subject = await _subjectCrawler.Crawl(int.Parse(courseId));
                 yield return await SubjectModel.CreateAsync(subject, _colorGenerator);
             }
         }
